Refactor visitors into their own header
