+++ dlib/cuda/cuda_dlib.h
@@ -6,7 +6,7 @@
 
 #include "tensor.h"
 #include "../geometry/rectangle.h"
-#include "../dnn/misc.h"
+#include "../dnn/utilities.h"
 
 namespace dlib
 {
